+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.1] - 2021-02-11
+
 ## Changed
 
 - cmd/ls: add demo message when running `ls` with no projects.
@@ -95,7 +97,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial pre-release.
 
-[Unreleased]: https://github.com/creekorful/srcode/compare/v0.7.0...HEAD
+[Unreleased]: https://github.com/creekorful/srcode/compare/v0.7.1...HEAD
+[0.7.1]: https://github.com/creekorful/srcode/compare/v0.7.1...HEAD
 [0.7.0]: https://github.com/creekorful/srcode/compare/v0.7.0...HEAD
 [0.6.0]: https://github.com/creekorful/srcode/compare/v0.6.0...HEAD
 [0.5.0]: https://github.com/creekorful/srcode/compare/v0.5.0...HEAD